Applied layouts extends and added new generic controller
Changed layouts to extends from app.blade.php instead replicate all
included code everywhere. Now we have sections in layouts.
Removed unused import on FireController relative to $request variable
and namespace.
Added genericController and updated web.php to represent that, this
controller are responsible by the index , about , info and manifest pages.
Fixed bug on main.js file relative to map init, before this the
script try to init a map in every page, now it will validate if we
have a element with the id map and if this is the case it will start the
map rendering.