Added default value for fake calls and alarms marker size
Added default value for  fake calls and alarms marker size to avoid
highlight this situations against more relevant ones.
Added missing titles to all pages so user can track the page we they are.
Changed pages layouts to use cards layout everywhere.
Changed styles for icon and fire status on list.blade.php since the styles
was related to the ones applied to map and most of the times the icon hided
the status label.